requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid==2.4.2
attrs==20.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
click==7.1.2
crypto==1.4.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Ext==0.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
idna==2.9
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
Naked==0.1.31
pika==1.1.0
Pillow==7.1.2
protobuf==3.13.0
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.9.7
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
pyotp==2.4.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scrypt==0.8.13
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.9
Werkzeug==0.16.1
wrapt==1.12.1
